NHDR-102 refactor: Naver Convention fromat(User)-세영
