+++ src/main/java/org/scoula/user/controller/UserController.java
@@ -1,12 +1,17 @@
 package org.scoula.user.controller;
 
-import lombok.RequiredArgsConstructor;
-import lombok.extern.log4j.Log4j2;
-
-import org.scoula.user.dto.UserDTO;
+import org.scoula.user.dto.UserDto;
 import org.scoula.user.service.UserService;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.log4j.Log4j2;
 
 @Log4j2
 @RestController
@@ -17,14 +22,14 @@ public class UserController {
 	private final UserService userService;
 
 	@PostMapping("/join")
-	public ResponseEntity<Void> join(@RequestBody UserDTO userDTO) {
-		userService.join(userDTO);
+	public ResponseEntity<Void> join(@RequestBody UserDto userDto) {
+		userService.join(userDto);
 		return ResponseEntity.ok().build();
 	}
 
 	@GetMapping("/{email}")
-	public ResponseEntity<UserDTO> getUser(@PathVariable String email) {
-		UserDTO user = userService.getUser(email);
+	public ResponseEntity<UserDto> getUser(@PathVariable String email) {
+		UserDto user = userService.getUser(email);
 		return ResponseEntity.ok(user);
 	}
 }
