+++ src/main/java/org/scoula/user/dto/AuthResultDto.java
@@ -7,7 +7,7 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class AuthResultDTO {
+public class AuthResultDto {
 	String token;
-	UserInfoDTO user;
+	UserInfoDto user;
 }
