+++ src/main/java/org/scoula/user/dto/UserInfoDto.java
@@ -1,20 +1,20 @@
 package org.scoula.user.dto;
 
+import org.scoula.user.domain.UserVo;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import org.scoula.user.domain.UserVO;
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class UserInfoDTO {
+public class UserInfoDto {
 	String email;
 	String userName;
 
-	public static UserInfoDTO of(UserVO userVO) {
-		return new UserInfoDTO(
+	public static UserInfoDto of(UserVo userVO) {
+		return new UserInfoDto(
 			userVO.getEmail(),
 			userVO.getUserName()
 		);
