+++ src/main/resources/org/scoula/user/mapper/UserMapper.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace="org.scoula.user.mapper.UserMapper">
-    <resultMap id="userMap" type="org.scoula.user.domain.UserVO">
+    <resultMap id="userMap" type="org.scoula.user.domain.UserVo">
         <id property="email" column="email"/>
         <result property="userName" column="user_name"/>
         <result property="userPhone" column="user_phone"/>
@@ -24,15 +24,22 @@
 
     <update id="update">
         UPDATE user
-        SET user_name = #{userName}, user_phone = #{userPhone}, birth = #{birth}, branch_name = #{branchName}, branch_branch_name = #{branchBranchName}
+        SET user_name          = #{userName},
+            user_phone         = #{userPhone},
+            birth              = #{birth},
+            branch_name        = #{branchName},
+            branch_branch_name = #{branchBranchName}
         WHERE email = #{email}
     </update>
 
     <delete id="delete">
-        DELETE FROM user WHERE email = #{email}
+        DELETE
+        FROM user
+        WHERE email = #{email}
     </delete>
 
     <select id="findAll" resultMap="userMap">
-        SELECT * FROM user
+        SELECT *
+        FROM user
     </select>
 </mapper>
